+++ CumulusMX/Stations/EcowittCloudStation.cs
@@ -261,11 +261,11 @@ namespace CumulusMX.Stations
 			StartLoop();
 		}
 
-		public override string GetEcowittCameraUrl()
+		public override string GetEcowittCameraUrl(int camIndex)
 		{
 			if (cumulus.ExtraSensorUseCamera ^ mainStation)
 			{
-				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress))
+				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
 				{
 					cumulus.LogWarningMessage("GetEcowittCameraUrl: Warning - URL requested, but no camera MAC address is configured");
 				}
@@ -273,8 +273,8 @@ namespace CumulusMX.Stations
 				{
 					try
 					{
-						EcowittCameraUrl = ecowittApi.GetCurrentCameraImageUrl(EcowittCameraUrl, Program.ExitSystemToken);
-						return EcowittCameraUrl;
+						EcowittCameraUrl[camIndex] = ecowittApi.GetCurrentCameraImageUrl(EcowittCameraUrl[camIndex], camIndex, Program.ExitSystemToken);
+						return EcowittCameraUrl[camIndex];
 					}
 					catch (Exception ex)
 					{
@@ -286,11 +286,11 @@ namespace CumulusMX.Stations
 			return string.Empty;
 		}
 
-		public override string GetEcowittVideoUrl()
+		public override string GetEcowittVideoUrl(int camIndex)
 		{
 			if (cumulus.ExtraSensorUseCamera ^ mainStation)
 			{
-				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress))
+				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
 				{
 					cumulus.LogWarningMessage("GetEcowittCameraUrl: Warning - URL requested, but no camera MAC address is configured");
 				}
@@ -298,8 +298,8 @@ namespace CumulusMX.Stations
 				{
 					try
 					{
-						EcowittVideoUrl = ecowittApi.GetLastCameraVideoUrl(EcowittVideoUrl, Program.ExitSystemToken);
-						return EcowittVideoUrl;
+						EcowittVideoUrl[camIndex] = ecowittApi.GetLastCameraVideoUrl(EcowittVideoUrl[camIndex], Program.ExitSystemToken);
+						return EcowittVideoUrl[camIndex];
 					}
 					catch (Exception ex)
 					{