+++ CumulusMX/Stations/HttpStationEcowitt.cs
@@ -281,11 +281,11 @@ namespace CumulusMX.Stations
 
 		}
 
-		public override string GetEcowittCameraUrl()
+		public override string GetEcowittCameraUrl(int camIndex)
 		{
 			if (cumulus.ExtraSensorUseCamera ^ mainStation)
 			{
-				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress))
+				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
 				{
 					cumulus.LogMessage("GetEcowittCameraUrl: Warning - URL requested, but no camera MAC address is configured");
 				}
@@ -293,8 +293,8 @@ namespace CumulusMX.Stations
 				{
 					try
 					{
-						EcowittCameraUrl = ecowittApi.GetCurrentCameraImageUrl(EcowittCameraUrl, Program.ExitSystemToken);
-						return EcowittCameraUrl;
+						EcowittCameraUrl[camIndex] = ecowittApi.GetCurrentCameraImageUrl(EcowittCameraUrl[camIndex], camIndex, Program.ExitSystemToken);
+						return EcowittCameraUrl[camIndex];
 					}
 					catch (Exception ex)
 					{
@@ -306,11 +306,11 @@ namespace CumulusMX.Stations
 			return string.Empty;
 		}
 
-		public override string GetEcowittVideoUrl()
+		public override string GetEcowittVideoUrl(int camIndex)
 		{
 			if (cumulus.ExtraSensorUseCamera ^ mainStation)
 			{
-				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress))
+				if (string.IsNullOrEmpty(cumulus.EcowittCameraMacAddress[camIndex]))
 				{
 					cumulus.LogMessage("GetEcowittVideoUrl: Warning - URL requested, but no camera MAC address is configured");
 				}
@@ -318,8 +318,8 @@ namespace CumulusMX.Stations
 				{
 					try
 					{
-						EcowittVideoUrl = ecowittApi.GetLastCameraVideoUrl(EcowittVideoUrl, Program.ExitSystemToken);
-						return EcowittVideoUrl;
+						EcowittVideoUrl[camIndex] = ecowittApi.GetLastCameraVideoUrl(EcowittVideoUrl[camIndex], camIndex, Program.ExitSystemToken);
+						return EcowittVideoUrl[camIndex];
 					}
 					catch (Exception ex)
 					{